Loan eligibility prediction function
Loan eligibility prediction model uses analysis techniques using the python packages like pandas, numpy, and algorithms like - Naive bayes and Decision Tree that modifies historical and present credit user information to create predictions whether the user is applicable to get loan or not. The final prediction is based on Naive bayes algorithm since the accuracy of the trained data on Naive bayes algorithm was 82% compared to Decision Tree algorithm which was 70% . Dataset was collected from Kaggle.